여러가지 수정

+++ src/main/java/com/daisobook/shop/booksearch/BooksSearch/dto/response/BookRespDTO.java
@@ -2,6 +2,7 @@ package com.daisobook.shop.booksearch.BooksSearch.dto.response;
 
 import com.daisobook.shop.booksearch.BooksSearch.entity.book.Status;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -14,9 +15,9 @@ public record BookRespDTO(
         List<AuthorRespDTO> authorList,
         String publisher,
         LocalDate publicationDate,
-        Integer price,
-        Integer discountPercentage,
-        Integer discountPrice,
+        Long price,
+        BigDecimal discountPercentage,
+        Long discountPrice,
         boolean isPackaging,
         Integer stock,
         Status status,
